[LEIP-83] Move hasher to libs/api
