+++ executables/collector/src/main/java/de/cyface/collector/handler/UserCreationHandler.java
@@ -22,7 +22,7 @@ import org.apache.commons.lang3.Validate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import de.cyface.collector.Hasher;
+import de.cyface.api.Hasher;
 import io.vertx.core.Handler;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
